+++ src/config.h
@@ -143,6 +143,7 @@ typedef struct {
   wstring appmouse_pointer;
   wstring pixmouse_pointer;
   int touch_scroll;
+  int drop_focus;
   // Selection
   bool input_clears_selection;
   bool copy_on_select;